src/main/java/faang/school/projectservice/filter/ProjectNameFilter.java
package faang.school.projectservice.filter;

import faang.school.projectservice.dto.filter.ProjectFilterDto;
import faang.school.projectservice.model.Project;
import org.springframework.stereotype.Component;

import java.util.regex.Pattern;
import java.util.stream.Stream;

@Component
public class ProjectNameFilter implements ProjectFilter {

    @Override
    public boolean isApplicable(ProjectFilterDto filters) {
        return filters.namePattern() != null;
    }

    @Override
    public Stream<Project> apply(Stream<Project> projects, ProjectFilterDto filters) {
        var pattern = Pattern.compile(Pattern.quote(filters.namePattern()), Pattern.CASE_INSENSITIVE);
        return projects.filter(project -> pattern.matcher(project.getName()).find());
    }
}
